requirements.txt
slack-sdk==3.26.1
python-dotenv==1.0.0
line-bot-sdk==3.7.0
openai==1.12.0 